feat: add favicon
Added the little icon thingy that shows up next to the title of a website. Used https://favicon.io/favicon-converter/ to convert the existing SE! logo to the file formats needed for the favicon to look nice across platforms. (Most of them are probably unnecessary, but good to have for browser compatibility reasons.)